Drop broken "@ symbol as icon" setting
It was merged into pleroma-fe on 2022-02-03 in
76547fe66d1771f5bff732a34b0547f890f4621a and imported
into akkoma-fe on 2022-06-08 with the merge commit
f6cf509a046507eeb9e52e2a279e82dd7a0f6bf8.

However, something went wrong in the merge and while the setting
and its infrastructure exist, it is never used anywhere and @ is
always displayed as text.

Given it existed in this broken state for nearly one and a half years,
never worked on akkoma-fe and no bugs were filed about this, it appears
nobody cares, so let’s just remove it.